refactor: generate user id msg function
